ux: relax case-sensitivity for login and auth operations
Removed case-sensitivity notes and updated UI for login, password reset, and username recovery to treat usernames and emails as case-insensitive. Updated changelog and bumped version to 1.8.5. Also added techdebt docs folder to .gitignore.

Most of the actual changes are in the API. I am backwards compatible with users that have accounts with the same username other than casing, but prevent more of these moving forward.